+++ src/material/autocomplete/testing/BUILD.bazel
@@ -15,21 +15,6 @@ ts_library(
     ],
 )
 
-ng_test_library(
-    name = "harness_tests_lib",
-    srcs = ["shared.spec.ts"],
-    deps = [
-        ":testing",
-        "//src/cdk/overlay",
-        "//src/cdk/testing",
-        "//src/cdk/testing/private",
-        "//src/cdk/testing/testbed",
-        "//src/material/autocomplete",
-        "@npm//@angular/forms",
-        "@npm//@angular/platform-browser",
-    ],
-)
-
 filegroup(
     name = "source-files",
     srcs = glob(["**/*.ts"]),
@@ -37,14 +22,14 @@ filegroup(
 
 ng_test_library(
     name = "unit_tests_lib",
-    srcs = glob(
-        ["**/*.spec.ts"],
-        exclude = ["shared.spec.ts"],
-    ),
+    srcs = glob(["**/*.spec.ts"]),
     deps = [
         ":testing",
+        "//src/cdk/testing",
+        "//src/cdk/testing/testbed",
         "//src/material/autocomplete",
-        "//src/material/autocomplete/testing:harness_tests_lib",
+        "@npm//@angular/forms",
+        "@npm//@angular/platform-browser",
     ],
 )
